add retries, error handling, and session expired
Three new functions on the session object to handle expired session tokens
and errors. setToken will update the token, onSessionExpired will be
called when the api indicates indicates an expired session.
onError will be called on all errors including sessionExpired.

The retry functionality implemented will retry on requests that failed due
to transport errors or server errors (5xx). The retries are less frequent
if the server continues to answer with errors. The retries will stop after
a 100 tries.