add retries, error and session-expired subscriptions
Added retries with backoff functionality on server and transport errors.
The retries will try 100 times with an increasing timeout up to 15
minutes.
Two new subscriptions:
error
This will listen to all errors that occurs in the sdk
session-expired
This will listen to session expired and by using setToken it is possible
to update the token.