Adds get-stage script
The get-stage script can be used to check if the project is in
pre-release or release stage.

A project is considered to be in release stage if its version is a
non-prerelease version and none of its dependencies can resolve to a
prerelease version.

A project is considered to be in prerelease stage if its version is
a prerelease version.