Removed requestAnimationFrame
After some discussion we decided to remove requestAnimationFrame because of timing issues when the focus
is removed from a tab or window. The effect can be that an update is stuck until the tab or window is active again.
This can happen to any update with the possibility of missing for example an alert because you switched windows.